refactor: TransactionManager에 롤백 로직이 담기도록 리팩터링
