+++ jdbc/src/main/java/com/interface21/jdbc/datasource/ConnectionManager.java
@@ -1,0 +1,24 @@
+package com.interface21.jdbc.datasource;
+
+import com.interface21.dao.DataAccessException;
+import java.sql.Connection;
+import java.sql.SQLException;
+import javax.sql.DataSource;
+
+public class ConnectionManager {
+
+    private final DataSource dataSource;
+
+    public ConnectionManager(final DataSource dataSource) {
+        this.dataSource = dataSource;
+    }
+
+    public Connection getConnection() {
+        try {
+            return dataSource.getConnection();
+        } catch (SQLException e) {
+            e.printStackTrace();
+            throw new DataAccessException("커넥션 연결 과정에서 문제가 발생했습니다");
+        }
+    }
+}
