+++ jdbc/src/main/java/com/interface21/jdbc/datasource/TransactionManager.java
@@ -5,38 +5,41 @@ import java.sql.Connection;
 import java.sql.SQLException;
 import java.util.function.Consumer;
 import java.util.function.Function;
-import javax.sql.DataSource;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 public class TransactionManager {
     private static final Logger logger = LoggerFactory.getLogger(TransactionManager.class);
 
-    private final DataSource dataSource;
+    private final ConnectionManager connectionManager;
 
-    public TransactionManager(DataSource dataSource) {
-        this.dataSource = dataSource;
+    public TransactionManager(ConnectionManager connectionManager) {
+        this.connectionManager = connectionManager;
     }
 
     public <T> T transaction(Function<Connection, T> function) {
-        try (Connection connection = dataSource.getConnection()) {
+        Connection connection = connectionManager.getConnection();
+        try (connection) {
             beginTransaction(connection);
             T result = function.apply(connection);
             endTransaction(connection);
             return result;
         } catch (Exception e) {
             logger.error(e.getMessage(), e);
+            rollback(connection);
             throw new DataAccessException(e);
         }
     }
 
     public void transaction(Consumer<Connection> consumer) {
-        try (Connection connection = dataSource.getConnection()) {
+        Connection connection = connectionManager.getConnection();
+        try (connection) {
             beginTransaction(connection);
             consumer.accept(connection);
             endTransaction(connection);
         } catch (Exception e) {
             logger.error(e.getMessage(), e);
+            rollback(connection);
             throw new DataAccessException(e);
         }
     }
@@ -48,5 +51,14 @@ public class TransactionManager {
     private void endTransaction(Connection connection) throws SQLException {
         connection.commit();
         connection.setAutoCommit(true);
+    }
+
+    private void rollback(Connection connection) {
+        try {
+            connection.rollback();
+            connection.setAutoCommit(true);
+        } catch (SQLException sqlException) {
+            throw new DataAccessException(sqlException);
+        }
     }
 }
